# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.19.1

package db

import (
	"database/sql/driver"
	"fmt"

	"github.com/jackc/pgx/v5/pgtype"
)

type CrawlStatus string

const (
	CrawlStatusUnknown       CrawlStatus = "unknown"
	CrawlStatusCompleted     CrawlStatus = "completed"
	CrawlStatusFailed        CrawlStatus = "failed"
	CrawlStatusBlocked       CrawlStatus = "blocked"
	CrawlStatusTimeout       CrawlStatus = "timeout"
	CrawlStatusInternalError CrawlStatus = "internal_error"
)

func (e *CrawlStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = CrawlStatus(s)
	case string:
		*e = CrawlStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for CrawlStatus: %T", src)
	}
	return nil
}

type NullCrawlStatus struct {
	CrawlStatus CrawlStatus
	Valid       bool // Valid is true if CrawlStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullCrawlStatus) Scan(value interface{}) error {
	if value == nil {
		ns.CrawlStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.CrawlStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullCrawlStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.CrawlStatus), nil
}

type ServerStatus string

const (
	ServerStatusUnknown ServerStatus = "unknown"
	ServerStatusOnline  ServerStatus = "online"
	ServerStatusOffline ServerStatus = "offline"
)

func (e *ServerStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = ServerStatus(s)
	case string:
		*e = ServerStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for ServerStatus: %T", src)
	}
	return nil
}

type NullServerStatus struct {
	ServerStatus ServerStatus
	Valid        bool // Valid is true if ServerStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullServerStatus) Scan(value interface{}) error {
	if value == nil {
		ns.ServerStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.ServerStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullServerStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.ServerStatus), nil
}

type Crawl struct {
	ID                pgtype.UUID
	ServerID          pgtype.UUID
	Status            CrawlStatus
	ErrorMsg          pgtype.Text
	StartedAt         pgtype.Timestamptz
	SoftwareName      pgtype.Text
	NumberOfPeers     pgtype.Int4
	OpenRegistrations pgtype.Bool
	TotalUsers        pgtype.Int4
	ActiveHalfYear    pgtype.Int4
	ActiveMonth       pgtype.Int4
	LocalPosts        pgtype.Int4
	LocalComments     pgtype.Int4
}

type PeeringRelationship struct {
	ServerID pgtype.UUID
	PeerID   pgtype.UUID
}

type Server struct {
	ID           pgtype.UUID
	Domain       string
	Status       ServerStatus
	CreatedAt    pgtype.Timestamptz
	DeletedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
	SoftwareName pgtype.Text
	LastCrawlID  pgtype.UUID
}